replace folium ecomap with lonboard
