+++ ecoscope/mapping/__init__.py
@@ -1,22 +1,5 @@
-from ecoscope.mapping.map import (
-    ControlElement,
-    EcoMap,
-    FloatElement,
-    NorthArrowElement,
-    ScaleElement,
-    GeoTIFFElement,
-    PrintControl,
-)
-
-from ecoscope.mapping.lonboard_map import EcoMap2
+from ecoscope.mapping.map import EcoMap
 
 __all__ = [
-    "ControlElement",
     "EcoMap",
-    "EcoMap2",
-    "FloatElement",
-    "NorthArrowElement",
-    "ScaleElement",
-    "GeoTIFFElement",
-    "PrintControl",
 ]
